Add myself as Emeritus
Hello!

I don't want to hold a such crucial spot! The ambassador program is limited in size, and it is important for all the members to be motivated. I am sure I will find other ways to serve the CNCF community moving forward!

it was a pleasure to be an ambassador!